+++ lib/hive_cache.dart
@@ -1,4 +1,3 @@
 library hive_cache;
 
-export 'src/hive_storage.dart';
 export 'src/hive_cache.dart';